third_party/WebKit/LayoutTests/http/tests/navigation/error404-frames.html

<html>
  <!-- Tests that 404s in frames are handled correctly -->
  <script>
    if (window.testRunner) {
        testRunner.dumpAsText();
        testRunner.dumpChildFrames();
        testRunner.clearBackForwardList();
        testRunner.dumpBackForwardList();        
    }
  </script>
  <frameset rows="90%,10%">
    <frame src="resources/error404.pl" name="main">
    <frame src="resources/otherpage.html" name="footer">
  </frameset>
</html>
