third_party/blink/web_tests/platform/generic/http/tests/navigation/error404-frames-expected.txt



--------
Frame: 'main'
--------
This 404 error was intentionally generated by a test script.

--------
Frame: 'footer'
--------
This is just a minimal page that we navigate to as part of testing back/forward.

============== Back Forward List ==============
curr->  http://127.0.0.1:8000/navigation/error404-frames.html
            http://127.0.0.1:8000/navigation/resources/otherpage.html (in frame "footer")
            http://127.0.0.1:8000/navigation/resources/error404.pl (in frame "main")
===============================================
